refactor: extract device registry helper methods to RegistryOperations
Move first batch of registry methods from main.py to coordinator/registry.py:
- _call_device_registry_api: Core registry call with compatibility handling
- _device_registry_kwargs_need_legacy_retry: Legacy kwarg detection
- _device_registry_build_legacy_kwargs: Legacy kwarg translation (static)
- _device_registry_config_subentry_kwarg_name: Subentry kwarg detection
- _device_registry_allows_translation_update: Translation support check

main.py: 8,356 → 8,227 lines (-129 lines)
registry.py: 0 → 177 lines

Phase 2 progress: 129 of ~1,682 lines extracted
